+++ Gems/Terrain/Code/Tests/TerrainHeightGradientListTests.cpp
@@ -16,6 +16,7 @@
 #include <GradientSignal/Ebuses/MockGradientRequestBus.h>
 #include <LmbrCentral/Shape/MockShapes.h>
 #include <Terrain/MockTerrainLayerSpawner.h>
+#include <Terrain/MockTerrain.h>
 #include <Tests/Mocks/Terrain/MockTerrainDataRequestBus.h>
 
 using ::testing::_;
