+++ Gems/Terrain/Code/terrain_mocks_files.cmake
@@ -9,4 +9,5 @@
 set(FILES
     Mocks/Terrain/MockTerrainLayerSpawner.h
     Mocks/Terrain/MockTerrainAreaSurfaceRequestBus.h
+    Mocks/Terrain/MockTerrain.h
 )
